+++ Blogs/Blogs/Screens/Cells/Utilities Cell/UtiliesCellViewModel.swift
@@ -6,35 +6,32 @@
 //
 
 import Foundation
-import UIKit
 
 final class UtiliesCellViewModel: UtiliesCellIdentifiable {
     
+    //MARK: Create Variable
     let utility: Utility
-    
     var cellIdentifier: String {
         "UtilitiesCollectionViewCell"
     }
     
-    var cellHeight: CGFloat {
+    var cellHeight: Float {
         44
     }
     
-    var cellWidth: CGFloat {
+    var cellWidth: Float {
        44
     }
 
-    
     init(utility: Utility) {
         self.utility = utility
     }
-    
 }
 
 final class UtiliesSectionViewModel: UtiliesSectionRowPresentable {
     var rows: [UtiliesCellIdentifiable] = []
     
-    var minimumLineSpacing: CGFloat {
+    var minimumLineSpacing: Float {
         4
     }
 }
